perf(metadata): add maintained indexes for validation
